LPS-118700 Move configuration panels to structure plugin directory
