+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -17,9 +17,6 @@
 @import './Translation';
 @import './layout-data-items/RowWithControls';
 
-@import './floating-toolbar/ContainerStylesPanel';
-@import './floating-toolbar/RowStylesPanels';
-
 @import './fragment-content/FragmentContent';
 
 @import './undo/UndoHistory';
