+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/Plugin.scss
@@ -1,3 +1,6 @@
+@import './floating-toolbar/ContainerStylesPanel';
+@import './floating-toolbar/RowStylesPanels';
+
 @import './ItemConfiguration';
 @import './PageStructureSidebar';
 @import './PageStructureSidebarSection';
