+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/plugins/page-structure/components/floating-toolbar/RowConfigurationPanel.test.js
@@ -16,11 +16,11 @@ import '@testing-library/jest-dom/extend-expect';
 import {cleanup, fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
-import {ResizeContextProvider} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/ResizeContext';
-import {RowConfigurationPanel} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/RowConfigurationPanel';
-import {StoreAPIContextProvider} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/store';
-import updateItemConfig from '../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
-import updateRowColumns from '../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns';
+import {ResizeContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/components/ResizeContext';
+import {StoreAPIContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/store/index';
+import updateItemConfig from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
+import updateRowColumns from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns';
+import {RowConfigurationPanel} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/floating-toolbar/RowConfigurationPanel';
 
 const ITEM_CONFIG = {
 	gutters: true,
@@ -71,7 +71,7 @@ const renderComponent = ({
 	);
 
 jest.mock(
-	'../../../../../src/main/resources/META-INF/resources/page_editor/app/config',
+	'../../../../../../../src/main/resources/META-INF/resources/page_editor/app/config',
 	() => ({
 		config: {
 			availableViewportSizes: {
@@ -83,12 +83,12 @@ jest.mock(
 );
 
 jest.mock(
-	'../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig',
+	'../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig',
 	() => jest.fn()
 );
 
 jest.mock(
-	'../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns',
+	'../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns',
 	() => jest.fn()
 );
